[BOOK-3] доработка менеджерского api и базовых эндпоинтов

+++ src/main/kotlin/ru/liljarn/booker/domain/model/dto/Genre.kt
@@ -1,6 +1,6 @@
 package ru.liljarn.booker.domain.model.dto
 
 data class Genre(
-    val genreId : Long,
+    val genreId : Int,
     val genreName: String
 )
