+++ src/main/kotlin/ru/liljarn/booker/domain/repository/GenreRepository.kt
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Repository
 import ru.liljarn.booker.domain.model.entity.GenreEntity
 
 @Repository
-interface GenreRepository : CrudRepository<GenreEntity, Long> {
-
+interface GenreRepository : CrudRepository<GenreEntity, Int> {
     fun findAllByGenreNameContainingIgnoreCase(genreName: String): List<GenreEntity>
 }
